requirements.txt
langchain==0.1.20
langchain-openai==0.1.8
openai==1.30.0
pandas==2.0.3
numpy==1.24.3
streamlit==1.28.0
weaviate-client==3.25.3
sentence-transformers==2.2.2
tavily-python==0.2.0
tqdm==4.65.0
python-dotenv==1.0.0
openpyxl==3.1.2
Pillow==10.0.0